Enforce UTF-8 encoding for DB accesses
For consistent writing to and reading from the database, we now enforce
UTF-8 encoding when opening the database connection.

Previously, the reading process was flawed and the result was latin1
encoding leading to bugs in further analysis.

To prevent problems with function implementations (see column 'impl' in
table 'commit_dependency'), we properly encode the content as UTF-8 now.

Signed-off-by: Raphael Nömmer <noemmer@fim.uni-passau.de>
Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>